requirements.txt
github3.py
click
